requirements.txt
fastapi
uvicorn
streamlit>=1.44.0
streamlit-folium>=0.23.0
earthengine-api>=0.1.400
google-auth>=2.28.0
torch>=2.9.0
torchvision>=0.20.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pydeck>=0.9.0
scikit-image>=0.22.0
rasterio>=1.3.0
Pillow>=10.0.0
folium>=0.16.0
